Set up Jest and write tests for first 3 problems
The tests in the future are meant to give inputs to functions, and to
test the results that are returned from those functions. For now
however, I wrote these three basic tests to define the parameters,
define the expected results, and test. This is meant as the beginning of
the writing of these tests, and as practice with Jest. I will next
expand on these tests to require the taking in of returned results from
actual functions (that have not been created yet).