Fix 4th test
When I originally made my 4th test for the previous commit, I
inadvertently wrote a caesarCipher function. I reviewed the guidelines
of TDD, and I realized that I wrote this test wrong. The idea is to
write simple tests that include inputs to be sent to a future function,
and expected results from that same function. These tests are not meant
to test every aspect of the future function, they are meant as guides
for building them in a TDD manner. This was an important lesson learned.
I commented out my original test block, and will use the concept for the
writing of the caesarCipher function, but I will avoid that kind of
mistake for the next problem.