Write caesarCipher function
I need to also add that there were typos with the 4th test. The expected
results for the last three out of four sub tests were meant to be a
(whether capitalized or not) instead of z. I am still getting used to
TDD and realize that tests should not have typos while attempting to
write the function to test.